requirements.txt
PyYAML
Jinja2
bottle
requests
